+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- `EnvironResolver` supports "generic" environment variables (such as `DATABASE_URL`)
+when resolving URIs.
 
 ## [0.0.3] - 2024-08-26
 ### Fixed